Additional refactoring for Async API
